% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_isolatedDrugNet.R
\name{drug_network_inference}
\alias{drug_network_inference}
\title{Generate isolated drug network}
\usage{
drug_network_inference(drug_input, drug_type, cpu = 1)
}
\arguments{
\item{drug_input}{A single-column dataframe containing the drug names.}

\item{drug_type}{The method used to construct the drug network. Using the "Virtual Nodes" strategy, each drug is linked to a virtual node.}

\item{cpu}{The number of cores to use for parallel processing.}
}
\value{
An unweighted isolated drug network, represented as a 3-column table (source = drug, dest = virtual node, weight = 1).
}
\description{
Generate isolated drug network
}
