% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coAbundanceNet.R
\name{gen_coAbundanceNet}
\alias{gen_coAbundanceNet}
\title{Generate co-abundance network}
\usage{
gen_coAbundanceNet(omics_matrix, correction_method = NULL, cpu = 1, pth = 0.05)
}
\arguments{
\item{omics_matrix}{A matrix of dimensions genes X samples representing proteomics data.}

\item{correction_method}{The method used to correct the p-value of the statistical test (either "bonferroni" or "fdr").}

\item{cpu}{The number of cores to use for parallel processing.}

\item{pth}{A numeric value, ranging from 0 and 1, that will be applied to the p-value of Spearman Correlation test.}
}
\value{
An undirected co-expression network, represented as a 3-column table (source, dest, weight).
}
\description{
Generate co-abundance network
}
