% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceback.R
\name{gen_traceback_net}
\alias{gen_traceback_net}
\title{Traceback a node-node association}
\usage{
gen_traceback_net(traceback_RES)
}
\arguments{
\item{traceback_RES}{Output of 'traceback_link()'}
}
\value{
A network showing the links between the input nodes for each omics layer.
}
\description{
Traceback a node-node association
}
