% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sim_mat_MH.R
\name{generate_sim_mt_MH}
\alias{generate_sim_mt_MH}
\alias{gen_sim_mat_MH}
\title{Generate similarity matrix
Adapted from:
https://github.com/LPioL/MultiVERSE/blob/master/RWR/Functions_RWRMH.R
https://github.com/LPioL/MultiVERSE/blob/master/RWR/GenerateSimMatrix_MH.R}
\usage{
gen_sim_mat_MH(
  network1,
  network2,
  network_B,
  restart = 0.7,
  tau1 = NA,
  tau2 = NA,
  delta1 = 0.5,
  delta2 = 0.5,
  lambda = 0.5,
  layer_transition_1 = NULL,
  layer_transition_2 = NULL,
  jump_neighborhood_1 = FALSE,
  weighted_multiplex_1 = FALSE,
  jump_neighborhood_2 = FALSE,
  weighted_multiplex_2 = FALSE,
  aggregation_method = "Sum",
  get_completeRWRmat = FALSE,
  no_seed_nodes = NULL,
  cores = 1
)
}
\arguments{
\item{network1}{a biological network}

\item{network2}{a pharmacological network}

\item{network_B}{a bipartite network}

\item{restart}{A real in the range of 0-1, it is the probability to restart the algorithm in the starting point}

\item{tau1}{tau1}

\item{tau2}{tau2}

\item{delta1}{delta}

\item{delta2}{delta}

\item{lambda}{lambda}

\item{layer_transition_1}{cond_jump}

\item{layer_transition_2}{cond_jump}

\item{jump_neighborhood_1}{A boolean, if false, the nodes between omics will be connected if they are the same, if true, the nodes will be connected with itself and its neighborhood but in the other omics}

\item{weighted_multiplex_1}{A boolean, if true, the edge between omics will be weighted. It is considered only if jump_neighborhood is true}

\item{jump_neighborhood_2}{A boolean, if false, the nodes between omics will be connected if they are the same, if true, the nodes will be connected with itself and its neighborhood but in the other omics}

\item{weighted_multiplex_2}{A boolean, if true, the edge between omics will be weighted. It is considered only if jump_neighborhood is true}

\item{aggregation_method}{aggregation method}

\item{get_completeRWRmat}{boolean}

\item{no_seed_nodes}{vector of characters specifying the nodes to exclude from the restarting procedure (e.g. no_seed_nodes = c('afatinib', 'TP53', 'ADH1A'))}

\item{cores}{Number of threads for Parallelization. It has to be positive integer. If it is equal to 1, no parallelization is not performed}
}
\value{
RWRMH_similarity
}
\description{
Research group:  Léo Pio-Lopez, Alberto Valdeolivas, Laurent Tichit, Élisabeth Remy, Anaïs Baudot
Publication:     https://arxiv.org/abs/2008.10085
}
