% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_adjList}
\alias{get_adjList}
\title{Convert adjacency matrix to adjacency list}
\usage{
get_adjList(adj_mat)
}
\arguments{
\item{adj_mat}{A square adjacency matrix where rows and columns correspond to nodes.}
}
\value{
A tibble with three columns: \code{source}, \code{dest}, and \code{weight}.
}
\description{
This function takes an adjacency matrix and returns an adjacency list.
It removes the upper triangle, zeros, and self-loops.
}
