% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filtered_corMat_by_adj_pval.R
\name{get_filtered_corMat_by_adj_pval}
\alias{get_filtered_corMat_by_adj_pval}
\title{Function to infer correlation network filtered by corrected p-value}
\arguments{
\item{input_matrix}{A matrix of dimensions genes X samples.}

\item{cor_method}{The correlation method to use (e.g., "pearson", "spearman").}

\item{adj_method}{The method used to correct the p-value (either "bonferroni" or "fdr").}

\item{cpu}{The number of cores to use for parallel processing.}

\item{pth}{A numeric value, ranging from 0 and 1, that will be applied to the p-value of Pearson Correlation test.}
}
\value{
A gene X gene correlation matrix, filtered by a corrected p-value lower than 0.05.
}
\description{
Function to infer correlation network filtered by corrected p-value
}
