% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAnnotation.R
\name{loadAnnot}
\alias{loadAnnot}
\title{Load annotation file}
\usage{
loadAnnot(annot_type = "gene_drug")
}
\arguments{
\item{annot_type}{A character string defining the annotation file to upload. Accepted values are:
'BRCA_sample' to upload the clinical information of BRCA samples used in this study, and 'gene_drug'
to upload the comprehensive annotation table of known genes and drugs. This latter is generated by consulting different databases,
such as COSMIC, HGNC, and DrugBank.}
}
\value{
A dataframe containing the requested annotation data.
}
\description{
This function loads an annotation file included in the \code{MiDNE} package.
}
\examples{
# Load BRCA sample annotation file
BRCA_sample_annot <- loadAnnot(annot_type = 'BRCA_sample')

# Load gene-drug annotation table
gene_drug_annot <- loadAnnot(annot_type = 'gene_drug')
}
