% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDrugNet.R
\name{loadDrugNet}
\alias{loadDrugNet}
\title{Load table of drug IDs}
\usage{
loadDrugNet(net_type = "vn")
}
\arguments{
\item{net_type}{A character defining the drug network to upload. Accepted values are:
'vn' to upload the drug network generated with the "virtual nodes" startegy.}
}
\value{
A dataframe with three columns: the source column reports the drug IDs,
the second one the virtual nodes, and the third one the weights of the edges.
}
\description{
This function loads a table of drug IDs included in the \code{MiDNE} package.
}
