% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDrugs.R
\name{loadDrugs}
\alias{loadDrugs}
\title{Load table of drug IDs}
\usage{
loadDrugs(drug_type = "FDA")
}
\arguments{
\item{drug_type}{A character defining the drug table to upload. Accepted values are:
'all' to upload all drugs interacting with at least one protein in the DrugBank database, and 'FDA'
to upload only the pharmacologically active FDA-approved drugs.}
}
\value{
A dataframe with one column reporting drug IDs.
}
\description{
This function loads a table of drug IDs included in the \code{MiDNE} package.
}
\examples{
# Load all drugs interacting with proteins
all_drugs <- loadDrugs(drug_type = 'all')

# Load FDA-approved drugs
fda_drugs <- loadDrugs(drug_type = 'FDA')
}
