% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOmicsNet.R
\name{loadOmicsNet}
\alias{loadOmicsNet}
\title{Load TCGA-BRCA omics networks}
\usage{
loadOmicsNet(
  net_type = c("expression", "proteomics", "methylation", "del", "amp", "all")
)
}
\arguments{
\item{net_type}{A character vector containing the names of omics networks to upload. Accepted values are:
'expression', 'proteomics', 'methylation', 'amp', 'del' and 'all'. You can provide a combination of the first four values
(e.g., \code{c('expression', 'proteomics')}).}
}
\value{
A named list of TCGA-BRCA networks.
The names of the list correspond to the omics types provided in \code{net_type}.
}
\description{
This function loads specific omics networks included in the \code{MiDNE} package.
}
