% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceback.R
\name{plot_traceback}
\alias{plot_traceback}
\title{Plot traceback output}
\usage{
plot_traceback(traceback_RES)
}
\arguments{
\item{traceback_RES}{Output of 'traceback_link()'}
}
\value{
A ggplot object
}
\description{
Plot traceback output
}
