% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceback.R
\name{traceback_link}
\alias{traceback_link}
\title{Traceback a node-node association}
\usage{
traceback_link(extended_RWR, links_list, reverse = FALSE, cpu = 1)
}
\arguments{
\item{extended_RWR}{A proximity matrix generated with gen_sim_mat_M or gen_sim_mat_MH functions by setting 'get_completeRWRmat = TRUE'.}

\item{links_list}{A list of pairs of nodes (e.g., list(c('TP53', 'SOX2'), c('TP53', 'KLF5')))}

\item{reverse}{Logical, whether to include reverse direction (target --> source)}

\item{cpu}{The number of cores to use for parallel processing.}
}
\value{
A tibble with columns: pair, source, target, proximity, layer
}
\description{
Traceback a node-node association
}
