% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_test_post_hoc.R
\name{chisq_theo_multcomp}
\alias{chisq_theo_multcomp}
\title{Pairwise comparisons after a chi-squared test for given probabilities (adapted from RVAideMemoire::chisq.theo.multcomp)}
\usage{
chisq_theo_multcomp(x, p = rep(1/length(x), length(x)), p.method = "fdr")
}
\arguments{
\item{x}{a contigency table}

\item{p}{theoretical proportions}

\item{p.method}{method for p-values correction. See help of p.adjust.}
}
\value{
a table with the results of the pairwise comparisons
}
\description{
Performs pairwise comparisons after a global chi-squared test for given probabilities.
}
