% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_test_post_hoc.R
\name{fisher_test_post_hoc}
\alias{fisher_test_post_hoc}
\title{Fisher's Exact Test and post-hoc analysis}
\usage{
fisher_test_post_hoc(matrix, cpu, correction_method, pth = 0.05)
}
\arguments{
\item{matrix}{A matrix of dimensions genes X samples.}

\item{cpu}{The number of cores to use for parallel processing.}

\item{correction_method}{The method used to correct the p-value (either "bonferroni" or "fdr").}

\item{pth}{A numeric value, ranging from 0 and 1, that will be applied to the p-value of the Fisher's Exact Test and of post-hoc analysis.}
}
\value{
A gene X gene matrix.
}
\description{
Fisher's Exact Test and post-hoc analysis
}
