% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coCNVnet.R
\name{gen_coCNVnet}
\alias{gen_coCNVnet}
\title{Generate co-CNV network}
\usage{
gen_coCNVnet(omics_matrix, correction_method = NULL, cpu = 1)
}
\arguments{
\item{omics_matrix}{A matrix of dimensions genes X samples representing Copy Number Variation (CNV) data.}

\item{correction_method}{The method used to correct the p-value of the statistical test (either "bonferroni" or "fdr").}

\item{cpu}{The number of cores to use for parallel processing.}
}
\value{
A list containing two undirected networks (co-amplification and co-deletion), each represented as a 3-column table (source, dest, weight).
}
\description{
Generate co-CNV network
}
