% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coDNAmethNet.R
\name{gen_coDNAmethNet}
\alias{gen_coDNAmethNet}
\title{Generate co-methylation network}
\usage{
gen_coDNAmethNet(omics_matrix, correction_method = NULL, th = 0.3, cpu = 1)
}
\arguments{
\item{omics_matrix}{A matrix of dimensions genes X samples representing DNA methylation data (beta-values).}

\item{correction_method}{The method used to correct the p-value of the statistical test (either "bonferroni" or "fdr").}

\item{th}{A positive integer used to binarize the input DNA methylation matrix.}

\item{cpu}{The number of cores to use for parallel processing.}
}
\value{
A list containing two undirected networks (co-amplification and co-deletion), each defined as a 3-column table (source, dest, weight).
}
\description{
Generate co-methylation network
}
