% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coExpressionNet.R
\name{gen_coExpressionNet}
\alias{gen_coExpressionNet}
\title{Generate co-expression network}
\usage{
gen_coExpressionNet(omics_matrix, correction_method = NULL, cpu = 1)
}
\arguments{
\item{omics_matrix}{A matrix of dimensions genes X samples representing transcriptomics data.}

\item{correction_method}{The method used to correct the p-value of the statistical test (either "bonferroni" or "fdr").}

\item{cpu}{The number of cores to use for parallel processing.}
}
\value{
An undirected co-expression network, represented as a 3-column table (source, dest, weight).
}
\description{
Generate co-expression network
}
