% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBipartiteNet.R
\name{loadBipartiteNet}
\alias{loadBipartiteNet}
\title{Load gene-drug network}
\usage{
loadBipartiteNet()
}
\value{
A dataframe with three columns reporting genes as source nodes, drugs as target nodes and the weights of the edges.
}
\description{
This function loads gene-drug network included in the \code{MiDNE} package.
}
