% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOmicsNet.R
\name{loadOmicsNet}
\alias{loadOmicsNet}
\title{Load TCGA-BRCA omics matrices.}
\usage{
loadOmicsNet(
  net_type = c("expression", "proteomics", "methylation", "del", "amp", "all")
)
}
\arguments{
\item{net_type}{A character vector containing the names of omics networks to upload. Accepted values are:
'expression', 'proteomics', 'methylation', 'amp', 'del' and 'all'. You can provide a combination of the first four values
(e.g., \code{c('expression', 'proteomics')}).}
}
\value{
A named list of TCGA-BRCA matrices with genes on rows and samples on columns.
The names of the list correspond to the omics types provided in \code{net_type}.
}
\description{
This function loads specific omics matrices from the TCGA-BRCA dataset included in the \code{MiDNE} package.
}
\examples{
# Load expression and proteomics data
omics_data <- loadOmicsMat(net_type = c('expression', 'proteomics'))

# Load all data types
all_data <- loadOmicsMat(net_type = 'all')
}
