% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_network_inference.R
\name{omics_network_inference}
\alias{omics_network_inference}
\title{Generate co-expression network}
\usage{
omics_network_inference(
  omics_matrix,
  omics_type = NULL,
  correction_method = NULL,
  th = NA,
  cpu = 1
)
}
\arguments{
\item{omics_matrix}{a genesXsamples matrix.}

\item{omics_type}{character that indicates the molecular domain of the omics matrix (Transcriptomics, Proteomics, Genetics, Epigenetics).}

\item{correction_method}{method for correcting the p-value of the statistical test (bonferroni or fdr).}

\item{th}{a numeric value to binarize a DNA methylation matrix (expressed as beta-values).}

\item{cpu}{number of cores to work in parallel.}
}
\value{
an undirected biological network, defined as a 3-columns table (source, dest, weight).
}
\description{
Generate co-expression network
}
