% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2D_matrix.R
\name{plot_2D_matrix}
\alias{plot_2D_matrix}
\title{Plot bidimensional reduction of a matrix}
\usage{
plot_2D_matrix(
  coord,
  nodes_anno,
  id_name,
  id_anno_color = NA,
  id_anno_shape = NA,
  interactive = TRUE,
  wo_legend = FALSE,
  title = "",
  shapes = NULL,
  size = 5
)
}
\arguments{
\item{coord}{a data frame with x and y coordinates on rows and observations on columns}

\item{nodes_anno}{annotation data frame with observations on rows and annotations on columns}

\item{id_name}{name of the column in \emph{nodes_anno} containing unique sample ids. There has to be correspondence on \emph{coord} columns names}

\item{id_anno_color}{name of the column in \emph{nodes_anno} whose values are mapped to colors to observations}

\item{id_anno_shape}{name of the column in \emph{nodes_anno} whose values are mapped to shapes to observations}

\item{interactive}{logical, if TRUE returns a plotly interactive plot, otherwise a standard ggplot is returned}

\item{wo_legend}{logical, if TRUE returns the plot without legend}

\item{title}{title of the plot}

\item{shapes}{shapes list to be used for plotting, if id_anno_shape is not NA}

\item{size}{size of the points}
}
\value{
a ggplot or a plotly object
}
\description{
Plot bidimensional reduction of a matrix
}
