% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psignif}
\alias{psignif}
\title{Transform p-value significance in characters (adapted from RVAideMemoire)}
\usage{
psignif(p)
}
\arguments{
\item{p}{p-value}
}
\value{
a string
}
\description{
transform p-value significance in characters.
}
